0x02 functions advance 16  8nd trail

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -8,7 +8,7 @@
  */
 int main(void)
 {
-	long long int pri = 1, sec = 2, sum = 0, total_sum = 0;
+	long unsigned int pri = 1, sec = 2, sum = 0, total_sum = 0;
 
 	sum = sec;
 	while (sum <= 4000000)
